Decoupled FunDebugger and FunDebugInterpreter.
