+++ src/main/java/ru/spbau/mit/debugger/FunDebugger.kt
@@ -2,6 +2,7 @@ package ru.spbau.mit.debugger
 
 import kotlinx.coroutines.experimental.runBlocking
 import ru.spbau.mit.ast.FunAst
+import ru.spbau.mit.debugger.FunDebugInterpreterReceiver.ExecutionPauseSnapshot
 import ru.spbau.mit.interpreter.FunContext
 import ru.spbau.mit.interpreter.FunInterpreter
 import ru.spbau.mit.interpreter.FunInterpreter.InterpretationResult
@@ -13,13 +14,22 @@ import kotlin.coroutines.experimental.*
 class FunDebugger(
     private val debugOut: PrintStream,
     private val programOut: PrintStream = debugOut
-) {
+) : FunDebugInterpreterReceiver {
     private var loadedAst: FunAst? = null
 
     private val breakpointMap: MutableMap<Int, FunAst.Expression?> = hashMapOf()
 
     private var pauseSnapshot: ExecutionPauseSnapshot? = null
     private var result: InterpretationResult? = null
+
+    override fun interpretationFinishedWith(result: InterpretationResult) {
+        this.result = result
+        pauseSnapshot = null
+    }
+
+    override fun interpretationPausedWith(pauseSnapshot: ExecutionPauseSnapshot) {
+        this.pauseSnapshot = pauseSnapshot
+    }
 
     fun load(filename: String) {
         loadedAst = buildAstFrom(filename)
@@ -50,7 +60,7 @@ class FunDebugger(
     fun run() {
         requireAstLoaded()
         requireProgramNonRunning()
-        val debugInterpreter = FunDebugInterpreter()
+        val debugInterpreter = FunDebugInterpreter(this, breakpointMap, programOut)
         pauseSnapshot = ExecutionPauseSnapshot(
             0, FunContext(), createInitialContinuation {
                 debugInterpreter.interpretAstDebugMode(loadedAst!!)
@@ -121,80 +131,6 @@ class FunDebugger(
     private fun requireProgramNonRunning() {
         if (pauseSnapshot != null) {
             throw FunDebugException("Program is already running!")
-        }
-    }
-
-    data class ExecutionPauseSnapshot(
-        val lineNumber: Int,
-        val executionContext: FunContext,
-        val executionContinuation: Continuation<Unit>
-    )
-
-    /** A visitor for interpreting the [FunAst] nodes in DEBUG mode. */
-    private inner class FunDebugInterpreter(
-        private val context: FunContext = FunContext(setOf(PRINTLN_FUN_NAME))
-    ) : FunInterpreter(programOut, context) {
-        suspend fun interpretAstDebugMode(ast: FunAst) {
-            result = visit(ast.rootNode)
-            pauseSnapshot = null
-        }
-
-        suspend private fun processBreakpointsAt(lineNumber: Int) {
-            if (breakpointMap.contains(lineNumber)) {
-                val condition = breakpointMap[lineNumber]
-                if (condition != null) {
-                    val interpreter = FunInterpreter(printStream, context.copy())
-                    try {
-                        val interpretationResult = interpreter.visit(condition)
-                        val conditionHolds = intToBool(interpretationResult.value!!)
-                        if (!conditionHolds) {
-                            return
-                        }
-                    } catch (e: Exception) {
-                        printStream.println(
-                            "Could not interpret specified condition ${e.message}:.")
-                    }
-                }
-                suspendCoroutine<Unit> { continuation ->
-                    pauseSnapshot = ExecutionPauseSnapshot(lineNumber, context.copy(), continuation)
-                }
-            }
-        }
-
-        override suspend fun visitVariable(variable: FunAst.Variable): InterpretationResult {
-            processBreakpointsAt(variable.lineNumber)
-            return super.visitVariable(variable)
-        }
-
-        override suspend fun visitWhileBlock(whileBlock: FunAst.WhileBlock): InterpretationResult {
-            processBreakpointsAt(whileBlock.lineNumber)
-            return super.visitWhileBlock(whileBlock)
-        }
-
-        override suspend fun visitIfStatement(
-            ifStatement: FunAst.IfStatement
-        ): InterpretationResult {
-            processBreakpointsAt(ifStatement.lineNumber)
-            return super.visitIfStatement(ifStatement)
-        }
-
-        override suspend fun visitAssignment(assignment: FunAst.Assignment): InterpretationResult {
-            processBreakpointsAt(assignment.lineNumber)
-            return super.visitAssignment(assignment)
-        }
-
-        override suspend fun visitReturnStatement(
-            returnStatement: FunAst.ReturnStatement
-        ): InterpretationResult {
-            processBreakpointsAt(returnStatement.lineNumber)
-            return super.visitReturnStatement(returnStatement)
-        }
-
-        override suspend fun visitFunctionCall(
-            functionCall: FunAst.FunctionCall
-        ): InterpretationResult {
-            processBreakpointsAt(functionCall.lineNumber)
-            return super.visitFunctionCall(functionCall)
         }
     }
 }
